Add interface to get run and test deps
Since tests are identified as a single command, dependencies for tests
and run are implicitly assumed to be the same as for the Build step,
including the target binary. This is, however not correct because it
doesn't account for the fact that we may want to use  a different test
runner or some additional target that must be compiled before testing,
but not necessarily in the build step of the target.

With this purpose this change introduces an extendedTestInterface
and extendedRunInteface where the user may also specify a TestDeps() or
RunDeps() function to get the dependencies of the test/run step.