Add assertions to ensure that interfaces are kept compatible
Given that all targets are traversed and examined against the interfaces
they implement, it is not trivial to detect when a target no longer
implements an interface. In order to detect such situations in any
potential refactors when a new method is added to one of this interfaces
or when it is modified, let's add a few functions that can be used to
assert statically at compile time that the interfaces are still met.

This makes propagation of breaking changes significantly less error
prone and straight forward.